% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{c2i}
\alias{c2i}
\title{This function calculates the incremental claims triangle from a cumulative claims triangle}
\usage{
c2i(triangle, is.square = FALSE)
}
\arguments{
\item{triangle}{Numeric matrix (n x n). Cumulative claims triangle or claims square}

\item{is.square}{Logical. FALSE if only a triangle (i.e. entries with indices i+j <= n+1) is given. TRUE if a claims square is given}
}
\description{
This function calculates the incremental claims triangle from a cumulative claims triangle
}
