% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL.R
\name{CL}
\alias{CL}
\title{This function provides the Chain Ladder Method}
\usage{
CL(triangle, is.incremental = FALSE)
}
\arguments{
\item{triangle}{Numeric matrix (n x n) containing the claims triangle}

\item{is.incremental}{Logical. Indicates whether the claims triangle is cumulative (default) or incremental.}
}
\value{
The function returns a list with various results:

\itemize{
\item Predictions: Predictions of the chain ladder method (cumulative)
\item Reserves: Reserves of the chain ladder method
\item MSE: MSE[i] = mean squared error of prediction for accident year i; MSE[n+1] = mean squared error of prediction for the total reserve
}
}
\description{
This function provides the Chain Ladder Method
}
\examples{
CL(glrm_example1_C)
CL(glrm_example1_S, is.incremental = FALSE)
}
