% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterfaceGLRM.R
\name{GLRM}
\alias{GLRM}
\title{This function provides the results of the Loss Ratio and the Gen. Loss Ratio Method in the Extended Additive Model}
\usage{
GLRM(
  triangle,
  volumes,
  is.incremental = FALSE,
  K = NULL,
  c = NULL,
  eps_ext = NULL,
  eps_start = NULL,
  eps_min = 0.001,
  eps_max = 0.1,
  s_sq_ext = NULL,
  Iterations = 50,
  UseAllEstimators = TRUE,
  Weights_LR = "optimal",
  Calculate_MSE_GLR = FALSE,
  Calculate_MSE_LR = FALSE,
  NumberOfSimulations = 10000,
  RunAllSimulations = FALSE,
  IterationsInSimMSE = 20,
  Use_eps_ext_inSimMSE = FALSE,
  Use_s_sq_ext_inSimMSE = FALSE,
  Seed = NULL,
  UseRcpp = TRUE,
  UseRelativities_s_sq = TRUE,
  Relativities_s_sq = NULL,
  SetIndicesRelativities = NULL,
  adjust_v_hat_infty_for_small_eps_ext = TRUE,
  Bias_Correction_m_hat_infty = TRUE
)
}
\arguments{
\item{triangle}{Numeric matrix (n x n) containing the claims triangle}

\item{volumes}{Numeric vector of lenght n containing the volume estimates}

\item{is.incremental}{Logical. Indicates whether the claims triangle is cumulative (default) or incremental.}

\item{K}{Numeric vector containing the columns with lognormally distributed increments. If K is set to NULL, then the maximum possible set is used.}

\item{c}{Vector of lenght n-1 containing the relativities of the epsilon_i. If c = NULL (defalut) then c = (1,...,1) is used, i.e. all epsilon_i are equal. c is normalized such that max(c) = 1.}

\item{eps_ext}{Numeric. External estimator for epsilon. If eps_ext = NULL (default) then epsilon is estimated in the recursion.}

\item{eps_start}{Numeric. Start value for epsilon in the recursion.}

\item{eps_min}{Numeric. Lower bound for the estimator for epsilon.}

\item{eps_max}{Numeric. Upper bound for the estimator for epsilon.}

\item{s_sq_ext}{Numeric vector. Contains external estimators for the \eqn{s_k^2}. If s_sq_ext = NULL (default) then epsilon is estimated in the recursion. Not used if eps_ext = 0.}

\item{Iterations}{Integer containing the number of iterations for the recursion.}

\item{UseAllEstimators}{Logical. If TRUE then all estimators r_{i,(j,k)} for 1/v_i are used. If FALSE then only those with (j,k)=(0,0) or j>i.}

\item{Weights_LR}{Either one of the strings 'optimal' or 'canonical' or a square matrix containing the weights to be used.}

\item{Calculate_MSE_GLR}{Logical. If FALSE then the prediction error of the generalized loss ratio method is not calculated (better performance)}

\item{Calculate_MSE_LR}{Logical. If FALSE then the prediction error of the loss ratio method is not calculated}

\item{NumberOfSimulations}{Numeric. Number of simulations for simulation of standard error.}

\item{RunAllSimulations}{Logical. If TRUE then also the time consuming simulations are run.}

\item{IterationsInSimMSE}{Integer. Number of iterations in the recursive parameter estimation in each simulation (if RunAllSimulations = TRUE).}

\item{Use_eps_ext_inSimMSE}{Logical. If TRUE then eps_ext is used in the simulation of the MSE with unfixed weights. If FALSE then eps_ext = NULL is used in the simulation.}

\item{Use_s_sq_ext_inSimMSE}{Logical. If TRUE then s_sq_ext is used in the simulation of the MSE with unfixed weights. If FALSE then s_sq_ext = NULL is used in the simulation.}

\item{Seed}{Numeric. If Seed != NULL then the value is used as seed for the simulations.}

\item{UseRcpp}{If TRUE then fast Rcpp routines are used.}

\item{UseRelativities_s_sq}{Logical. If TRUE (default) then relativities are used for the estimators of \eqn{s_k^2}}

\item{Relativities_s_sq}{Numeric vector of lenght n which contains the relativities if UseRelativities_s_sq = TRUE. If NULL then the relativities from the estimators of the loss ratio method are used.}

\item{SetIndicesRelativities}{Numeric vector of lenght < n. Contains the columns k for which the relativities are applied. If NULL and UseRelativities_s_sq = TRUE, then SetIndicesRelativities is set to 1:(n-1).}

\item{adjust_v_hat_infty_for_small_eps_ext}{Logical. If TRUE then adjustment of revised volume estimates v_hat_infty is reduced if the adjustment is not plausible for a given eps_ext or eps_max.}

\item{Bias_Correction_m_hat_infty}{Logical. If TRUE (default) then the estimators m_hat_infty are corrected for bias (otherwise weights d are used instead of h)}
}
\value{
The function returns a list with various results:

\itemize{
\item Summary_GLR: Overview to the results of the generalized loss ratio method in the extended additive model
\item Summary_LR: Overview to the results of the loss ratio method in the extended additive model
\item Predictions_GLR: Predictions of the generalized loss ratio method (cumulative)
\item Reserves_GLR: Reserves of the generalized loss ratio method
\item Predictions_LR: Predictions of the generalized loss ratio method (cumulative)
\item Reserves_LR: Reserves of the loss ratio method
\item Param_eps_hat_infty: Estimator for eps from the recursion.
\item Param_m_hat_infty: Estimator for m from the recursion.
\item Param_s_hat_sq_infty: Estimators for \eqn{s_k^2} from the recursion.
\item Param_v_hat_infty: Estimator for the volumes from the recursion.
\item Weights_LR: Matrix containing the used weights for the loss ratio method.
\item Weights_LR_Type: String containing the type of the chosen weights.
\item Recursion_eps_hat: Estimator for \eqn{\epsilon} for each step \eqn{k} of the recursion.
\item Recursion_v_hat: Estimators for the volumes \eqn{v_i} for each step of the recursion.
\item Recursion_m_hat: Estimators for \eqn{m_k} for each step of the recursion.
\item Recursion_s_hat_sq: Estimators for \eqn{s_k^2} for each step of the recursion.
\item se_GLR_fixed_weights_calc: Calculated standard errors of the generalized loss ratio method with fixed weights \eqn{g} and \eqn{h}.
\item se_GLR_fixed_weights_sim: Simulated standard errors of the generalized loss ratio method with fixed weights \eqn{g} and \eqn{h}.
\item se_GLR_sim: Simulated standard errors of the generalized loss ratio method, where the weights \eqn{g} and \eqn{h} are estimated in each simulation.
\item se_LR_fixed_weights_calc: Calculated standard errors of the generalized loss ratio method with fixed weights \eqn{w}.
\item se_LR_fixed_weights_sim: Simulated standard errors of the generalized loss ratio method with fixed weights \eqn{w}.
\item se_LR_sim: Simulated standard errors of the generalized loss ratio method, where the weights \eqn{w} are calculated in each simulation (in case of canonical or optimal weights).
\item se_CL_sim: Simulated standard errors of the chain ladder method (simulation based on the extended additive model).
}
}
\description{
This function provides the results of the Loss Ratio and the Gen. Loss Ratio Method in the Extended Additive Model
}
\examples{
GLRM(glrm_example1_C, glrm_example1_v_hat)
GLRM(glrm_example1_S, glrm_example1_v_hat, is.incremental = TRUE, K = 1:2, UseAllEstimators = FALSE,
     eps_ext = 0.03, Calculate_MSE_LR = TRUE, Calculate_MSE_GLR = TRUE, Weights_LR = "canonical")
GLRM(glrm_example2_C, glrm_example2_v_hat, Iterations = 100, RunAllSimulations = TRUE,
     NumberOfSimulations = 1000)
}
