% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{glrm_example2_C}
\alias{glrm_example2_C}
\title{Example data: Cumulative claims triangle (n=8)}
\format{An object of class \code{matrix} with 8 rows and 8 columns.}
\usage{
glrm_example2_C
}
\description{
Example data: Cumulative claims triangle (n=8)
}
\keyword{datasets}
