% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{glrm_example3_C}
\alias{glrm_example3_C}
\title{Example data: Cumulative claims triangle (n=15)}
\format{An object of class \code{matrix} with 15 rows and 15 columns.}
\usage{
glrm_example3_C
}
\description{
Example data: Cumulative claims triangle (n=15)
}
\keyword{datasets}
