% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{glrm_example3_S}
\alias{glrm_example3_S}
\title{Example data: Incremental claims triangle (n=15)}
\format{An object of class \code{matrix} with 15 rows and 15 columns.}
\usage{
glrm_example3_S
}
\description{
Example data: Incremental claims triangle (n=15)
}
\keyword{datasets}
