% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{glrm_example3_v_hat}
\alias{glrm_example3_v_hat}
\title{Example data: Volume estimates (n=15)}
\format{An object of class \code{numeric} of length 15.}
\usage{
glrm_example3_v_hat
}
\description{
Example data: Volume estimates (n=15)
}
\keyword{datasets}
