% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{i2c}
\alias{i2c}
\title{This function calculates the cumulative claims triangle from an incremental claims triangle}
\usage{
i2c(triangle, is.square = FALSE)
}
\arguments{
\item{triangle}{Numeric matrix (n x n). Incremental claims triangle or claims square}

\item{is.square}{Logical. FALSE if only a triangle (i.e. entries with indices i+j <= n+1) is given. TRUE if a claims square is given}
}
\description{
This function calculates the cumulative claims triangle from an incremental claims triangle
}
